small cleanup, mostly nullptr

+++ include/rgbd/shared_mem_client.h
@@ -32,7 +32,7 @@ private:
     boost::interprocess::mapped_region mem_buffer_header;
     boost::interprocess::mapped_region mem_image;
 
-    int sequence_nr;
+    uint64_t sequence_nr;
 
     BufferHeader* buffer_header;
